Minor app improvements
Added code to ensure that when the user taps on an item in the recycler view of the main activity, the scroll view will scroll if the view is not entirely visible to ensure that all data is displayed.  Also added some more metric to imperial conersions to the UtilityMethods class.